Init lenght more precise

+++ resources/lib/services/msl/converter.py
@@ -23,9 +23,9 @@ def convert_to_dash(manifest):
     has_drm_streams = manifest['hasDrmStreams']
     protection_info = _get_protection_info(manifest) if has_drm_streams else None
 
-    seconds = int(manifest['duration'] / 1000)
+    seconds = manifest['duration'] / 1000
     init_length = int(seconds / 2 * 12 + 20 * 1000)
-    duration = "PT" + str(seconds) + ".00S"
+    duration = "PT" + str(int(seconds)) + ".00S"
 
     root = _mpd_manifest_root(duration)
     period = ET.SubElement(root, 'Period', start='PT0S', duration=duration)
